fix the circular dependencies issues which related to the item store.. reorg all store effect to prevent other similar issues
