+++ src/app/community/store/effects.ts
@@ -1,0 +1,2 @@
+export * as communityEffects from './community.effects';
+export * as communityWsSubscriptionEffects from './community-ws-subscription.effects';
