+++ src/app/community/store/index.ts
@@ -1,4 +1,2 @@
 export { fromCommunity } from './community.store';
 export { communityPollActions, communityPageActions, communityActivityFeedActions } from './community.actions';
-export * as communityEffects from './community.effects';
-export * as communityWsSubscriptionEffects from './community-ws-subscription.effects';
