+++ src/app/items/containers/group-progress-grid/group-progress-grid.component.ts
@@ -21,7 +21,8 @@ import { mapToFetchState, readyData } from 'src/app/utils/operators/state';
 import { FetchState } from 'src/app/utils/state';
 import { HttpErrorResponse } from '@angular/common/http';
 import { allowsGivingPermToItem, ItemCorePerm } from 'src/app/items/models/item-permissions';
-import { itemRoute, itemRouteWith, selectObservedGroupRouteAsItemRouteParameter } from 'src/app/models/routing/item-route';
+import { itemRoute, itemRouteWith } from 'src/app/models/routing/item-route';
+import { selectObservedGroupRouteAsItemRouteParameter } from 'src/app/models/routing/item-route-observation-selector';
 import { Store } from '@ngrx/store';
 import { ItemRouter } from 'src/app/models/routing/item-router';
 import { GroupLinkPipe } from 'src/app/pipes/groupLink';
