+++ src/app/models/routing/item-route.ts
@@ -1,9 +1,6 @@
 import { ContentRoute } from './content-route';
 import { ItemTypeCategory } from '../../items/models/item-type';
 import { AnswerId, AttemptId, GroupId, ItemId, ItemPath, ParticipantId } from '../ids';
-import { createSelector } from '@ngrx/store';
-import { fromObservation } from 'src/app/store/observation';
-import { isUser } from './group-route';
 
 /* **********************************************************************************************************
  * Item Route: Object storing information required to navigate to an item
@@ -88,10 +85,3 @@ export function parentRoute(route: ItemRoute, defaultActivityRoute: FullItemRout
   });
 }
 
-/**
- * Selector to get the currently observed group as an item route parameter
- */
-export const selectObservedGroupRouteAsItemRouteParameter = createSelector(
-  fromObservation.selectObservedGroupRoute,
-  route => (route !== null ? { observedGroup: { id: route.id, isUser: isUser(route) } } : {})
-);
