+++ src/app/store/config/effects.ts
@@ -1,0 +1,1 @@
+export * as configEffects from './config.effects';
