+++ src/app/store/config/index.ts
@@ -1,3 +1,2 @@
 export { fromConfig } from './config.store';
-export * as configEffects from './config.effects';
 export { configActions } from './config.actions';
