+++ src/app/store/navigation/effects.ts
@@ -1,0 +1,2 @@
+export * as selectedContentEffects from './selected-content/selected-content.effects';
+export * as analyticsTrackingEffects from './analytics-tracking.effects';
