+++ src/app/store/navigation/index.ts
@@ -1,3 +1,1 @@
-export * as selectedContentEffects from './selected-content/selected-content.effects';
-export * as analyticsTrackingEffects from './analytics-tracking.effects';
 export { fromSelectedContent } from './selected-content/selected-content.store';
