+++ src/app/store/notification/effects.ts
@@ -1,0 +1,3 @@
+export * as notificationEffects from './notification.effects';
+export * as notificationWebsocketEffects from './notification-websocket.effects';
+export * as notificationThreadCleanupEffects from './notification-thread-cleanup.effects';
