+++ src/app/store/notification/index.ts
@@ -1,5 +1,2 @@
 export { fromNotification } from './notification.store';
 export { notificationApiActions, notificationWebsocketActions } from './notification.actions';
-export * as notificationEffects from './notification.effects';
-export * as notificationWebsocketEffects from './notification-websocket.effects';
-export * as notificationThreadCleanupEffects from './notification-thread-cleanup.effects';
