+++ src/app/store/websocket/effects.ts
@@ -1,0 +1,1 @@
+export * as websocketEffects from './websocket.effects';
