+++ src/app/store/websocket/index.ts
@@ -1,3 +1,2 @@
 export { websocketClientActions } from './websocket.actions';
 export { fromWebsocket, reducer, State, initialState } from './websocket.store';
-export * as websocketEffects from './websocket.effects';
